+++ drawHUD.h
@@ -5,6 +5,6 @@
 
 // Explicitly use std::function<void()> instead of auto
 std::function<void()> drawText(std::string text);
-std::function<void()> draw_ortho_compass(int rot_x);
+std::function<void()> draw_ortho_compass(float rot_x);
 
 void draw_HUD(const std::function<void()>& drawUI);
